i did something but at the same time i didn't really do anything also ignore the absolute path I left in there Keren told me to commit it before I got the chance to get the relative imports working so if I get caught please don't clown me thank you - oh and yeah javascript importing is somewhat working now with svelte components but still very mess
